requirements.txt
django==5.1.4
django-environ
psycopg2  # PostgreSQL用
django-storages  # S3などのストレージ用
gunicorn  # 本番環境でのWSGIサーバー
